Display Chat title in SimpleMessageDelegate if sender is chat
On overviewPage I found it sufficient being empty, since the title is directly above it, anyway. But in the channel itself it looked wrong without a user name.

The name link doesn't have a href – this would only be useful for edge cases like someone sharing a "changed title" message to another chat. I did not consider that relevant enough.